+++ c/column_int.cc
@@ -8,7 +8,6 @@
 #include "column.h"
 #include "datatablemodule.h"
 #include "python/int.h"
-#include "utils/parallel.h"