+++ c/csv/reader.cc
@@ -12,10 +12,10 @@
 #include "csv/reader.h"
 #include "csv/reader_arff.h"
 #include "csv/reader_fread.h"
+#include "parallel/api.h"
 #include "python/_all.h"
 #include "python/string.h"
 #include "utils/exceptions.h"
-#include "utils/parallel.h"
 #include "utils/misc.h"         // wallclock
 #include "datatable.h"
 #include "encodings.h"
@@ -100,21 +100,19 @@ void GenericReader::init_verbose() {
 }
 
 void GenericReader::init_nthreads() {
-  #ifdef DTNOOPENMP
-    nthreads = 1;
-    trace("Using 1 thread because datatable was built without OMP support");
-  #else
-    int32_t nth = freader.get_attr("nthreads").to_int32();
-    if (ISNA<int32_t>(nth)) {
-      nthreads = config::nthreads;
-      trace("Using default %d thread%s", nthreads, (nthreads==1? "" : "s"));
-    } else {
-      nthreads = config::normalize_nthreads(nth);
-      int maxth = config::normalize_nthreads(0);
-      trace("Using %d thread%s (requested=%d, max.available=%d)",
-            nthreads, (nthreads==1? "" : "s"), nth, maxth);
-    }
-  #endif
+  int32_t nth = freader.get_attr("nthreads").to_int32();
+  int maxth = static_cast<int>(dt::num_threads_in_pool());
+  if (ISNA<int32_t>(nth)) {
+    nthreads = maxth;
+    trace("Using default %d thread%s", nthreads, (nthreads==1? "" : "s"));
+  } else {
+    nthreads = nth;
+    if (nthreads > maxth) nthreads = maxth;
+    if (nthreads <= 0) nthreads += maxth;
+    if (nthreads <= 0) nthreads = 1;
+    trace("Using %d thread%s (requested=%d, max.available=%d)",
+          nthreads, (nthreads==1? "" : "s"), nth, maxth);
+  }
 }
 
 void GenericReader::init_fill() {
@@ -355,7 +353,8 @@ void GenericReader::_message(
     #pragma GCC diagnostic pop
   }
 
-  if (omp_get_thread_num() == 0) {
+  size_t ith = dt::this_thread_index();
+  if (ith + 1 <= 1) {
     try {
       Py_ssize_t len = static_cast<Py_ssize_t>(strlen(msg));
       PyObject* pymsg = PyUnicode_Decode(msg, len, "utf-8",
@@ -376,12 +375,10 @@ void GenericReader::_message(
 }
 
 void GenericReader::progress(double progress, int statuscode) {
-  xassert(omp_get_thread_num() == 0);
   freader.invoke("_progress", "(di)", progress, statuscode);
 }
 
 void GenericReader::emit_delayed_messages() {
-  xassert(omp_get_thread_num() == 0);
   if (delayed_message.size()) {
     trace("%s", delayed_message.c_str());
     delayed_message.clear();