+++ c/models/dt_ftrl.h
@@ -21,12 +21,9 @@
 //------------------------------------------------------------------------------
 #ifndef dt_MODELS_FTRL_h
 #define dt_MODELS_FTRL_h
-#include "str/py_str.h"
-#include "utils/parallel.h"
-#include "options.h"
 #include "models/utils.h"
-
-
+#include "str/py_str.h"
+#include "options.h"
 namespace dt {