+++ c/parallel/api.h
@@ -80,6 +80,16 @@ void parallel_region(function<void()> f);
 
 
 /**
+ * Make all threads within the current team wait until all threads have
+ * arrived to this point.
+ * Note: it is the user's responsibility to ensure that all threads CAN arrive
+ * at the barrier. If not, a deadlock will occur as the threads will be waiting
+ * for all the team to arrive before they could proceed.
+ */
+void barrier();
+
+
+/**
  * Run parallel loop `for i in range(nrows): f(i)`, with static scheduling.
  */
 template <typename F>
@@ -109,7 +119,7 @@ void parallel_for_dynamic(size_t nrows, function<void(size_t)> fn);
 /**
  * Execute loop
  *
- *     for i in range(nrows):
+ *     for i in range(n_iterations):
  *         pre-ordered(i)
  *         ordered(i)
  *         post-ordered(i)
@@ -127,9 +137,10 @@ class ordered {
     void parallel(function<void(size_t)> pre_ordered,
                   function<void(size_t)> do_ordered,
                   function<void(size_t)> post_ordered);
+    void set_n_iterations(size_t n);
 };
 
-void parallel_for_ordered(size_t nrows, size_t nthreads,
+void parallel_for_ordered(size_t n_iterations, size_t n_threads,
                           function<void(ordered*)> fn);