+++ ci/setup_utils.py
@@ -118,8 +118,6 @@ def ismacos():
 def iswindows():
     return sys.platform == "win32"
 
-def omp_enabled():
-    return not os.environ.get("DTNOOMP")
 
 
 
@@ -324,8 +322,6 @@ def get_compiler():
                     cc += ".exe"
                 try:
                     cmd = [cc, "-c", fname, "-o", outname]
-                    if omp_enabled():
-                        cmd += ["-fopenmp"]
                     proc = subprocess.Popen(cmd, stdout=subprocess.PIPE,
                                             stderr=subprocess.PIPE)
                     stdout, stderr = proc.communicate()
@@ -335,8 +331,6 @@ def get_compiler():
                         log.info("Compiler `%s` will be used" % cc)
                         print_compiler_version(log, cc)
                         return cc
-                    elif omp_enabled() and "-fopenmp" in stderr:
-                        log.info("Compiler `%s` does not support OpenMP" % cc)
                     else:
                         log.info("Compiler `%s` returned an error when "
                                  "compiling a blank file: <%s>" % (cc, stderr))
@@ -437,10 +431,6 @@ def get_extra_compile_flags():
         # Path to source files / Python include files
         flags += ["-Ic"]
 
-        # Enable OpenMP support
-        if omp_enabled():
-            flags.insert(0, "-fopenmp")
-
         # Generate 'Position-independent code'. This is required for any
         # dynamically-linked library.
         flags += ["-fPIC"]
@@ -540,9 +530,6 @@ def get_extra_link_args():
     with TaskContext("Determine the extra linker flags") as log:
         flags += ["-Wl,-rpath,%s" % get_rpath()]
 
-        if omp_enabled():
-            flags += ["-fopenmp"]
-
         # Omit all symbol information from the output
         # ld warns that this option is obsolete and is ignored. However with
         # this option the size of the executable is ~25% smaller...
@@ -578,7 +565,6 @@ def required_link_libraries():
     #   linux-vdso.so.1
     #   /lib64/ld-linux-x86-64.so.2
     #   libstdc++.so.6  => /usr/lib/x86_64-linux-gnu/libstdc++.so.6
-    #   libgomp.so.1    => /usr/lib/x86_64-linux-gnu/libgomp.so.1
     #   libgcc_s.so.1   => /lib/x86_64-linux-gnu/libgcc_s.so.1
     #   libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0
     #   libc.so.6       => /lib/x86_64-linux-gnu/libc.so.6
@@ -590,7 +576,6 @@ def required_link_libraries():
         return []
 
     # Clang on MacOS links to the following libraries (`otool -L`):
-    #   @rpath/libomp.dylib
     #   @rpath/libc++.1.dylib
     #   /usr/lib/libSystem.B.dylib
     # In addition, `libc++abi.1.dylib` (or `libc++abi.dylib`) is referenced
@@ -599,11 +584,9 @@ def required_link_libraries():
     #
     if is_clang():
         if ismacos():
-            return ["libomp.dylib", "libc++.1.dylib", "libc++abi.dylib"]
+            return ["libc++.1.dylib", "libc++abi.dylib"]
         if islinux():
-            return ["libomp.so", "libc++.so.1", "libc++abi.so.1"]
-        if iswindows():
-            return ["libomp.dll"]
+            return ["libc++.so.1", "libc++abi.so.1"]
     return []