+++ datatable/internal.py
@@ -26,7 +26,6 @@ from .lib._datatable import (
     frame_column_rowindex,
     frame_integrity_check,
     get_thread_ids,
-    has_omp_support,
     in_debug_mode,
     RowIndex
 )
@@ -36,7 +35,6 @@ __all__ = [
     "frame_column_rowindex",
     "frame_integrity_check",
     "get_thread_ids",
-    "has_omp_support",
     "in_debug_mode",
     "RowIndex",
 ]